+++ packages/clients/portal.manifest.mjs
@@ -1,6 +1,6 @@
 export default {
   id: 'clients',
-  version: '0.4.5',
+  version: '0.4.6',
   source: '@nuxt-customer-portal/clients',
   dependsOn: ['core', 'ui'],
   schema: './server/db/schema/clients.ts',