+++ packages/invoice-products/CHANGELOG.md
@@ -1,5 +1,14 @@
 # Product invoices
 
+## 0.4.6
+
+### Patch Changes
+
+- Updated dependencies
+  - @nuxt-customer-portal/core@0.4.6
+  - @nuxt-customer-portal/invoices@0.4.6
+  - @nuxt-customer-portal/products@0.4.6
+
 ## 0.4.5
 
 ### Patch Changes