+++ packages/invoice-timesheets/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @nuxt-customer-portal/invoice-timesheets
 
+## 0.4.6
+
+### Patch Changes
+
+- Updated dependencies
+  - @nuxt-customer-portal/core@0.4.6
+  - @nuxt-customer-portal/invoices@0.4.6
+  - @nuxt-customer-portal/timesheets@0.4.6
+
 ## 0.4.5
 
 ### Patch Changes